+++ backend/Frametric.Application/DTOs/Analytics/DecadeCountDto.cs
@@ -1,3 +1,3 @@
 namespace Frametric.Application.DTOs.Analytics;
 
-public record DecadeCountDto(int Decade, int Count);
+public record DecadeCountDto(int Decade, long Count);
