+++ backend/Frametric.Application/DTOs/Analytics/WeeklyWatchesDto.cs
@@ -1,3 +1,3 @@
 namespace Frametric.Application.DTOs.Analytics;
 
-public record WeeklyWatchesDto(string DayOfWeek, int Count);
+public record WeeklyWatchesDto(string DayOfWeek, long Count);
